Replace calling GetClientRect with win32App.m_width and m_height win32App is singleton instance of class Win32Application

+++ sturdy-guacamole/Win32Application.cpp
@@ -6,7 +6,7 @@ sturdy_guacamole::Win32Application::Win32Application()
 	WNDCLASSEXW wc{ sizeof(wc), CS_CLASSDC, DefWindowProcW, 0L, 0L, GetModuleHandle(NULL), NULL, NULL, NULL, NULL, m_className, NULL };
 	::RegisterClassExW(&wc);
 	
-	RECT rect{ 0, 0, m_width, m_height };
+	RECT rect{ 0, 0, static_cast<LONG>(m_width), static_cast<LONG>(m_height) };
 	::AdjustWindowRectEx(&rect, WS_OVERLAPPEDWINDOW, false, WS_EX_OVERLAPPEDWINDOW);
 
 	int adjusted_width{ rect.right - rect.left };
